+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking changes
 
 - Updated dependency to `@animoca/ethereum-contracts-assets@2.0.0`.
+- Removed `TokenLaunchpadVouchersRedeemer.sol` to integrated it as generic `ERC1155VouchersRedeemer.sol` in `@animoca/ethereum-contracts-assets`.
 - Removed Pausable feature from TokenLaunchpadVouchers.
 
 ## 1.1.0